Use modal to input game arguments when variant is doable

+++ discord/dcmaid/utils.py
@@ -53,10 +53,12 @@ def trim(string: Optional[str]):
 async def remove_thinking(ctx: QuasiContext):
 	if isinstance(ctx, discord.ApplicationContext):
 		await ctx.defer()
-		await ctx.delete()
+		if ctx.message is not None:
+			await ctx.delete()
 	else:
 		await ctx.response.defer()
-		await ctx.delete_original_message()
+		if ctx.message is not None:
+			await ctx.delete_original_message()
 
 def get_subcommand(group: discord.SlashCommandGroup, name: str) -> Optional[Union[discord.SlashCommand, discord.SlashCommandGroup]]:
 	for cmd in group.subcommands:
